docs: update all public documents to use auto-generated admin clients.
