+++ samples/samples/backup_sample.py
@@ -19,35 +19,46 @@ For more information, see the README.rst under /spanner.
 """
 
 import argparse
-from datetime import datetime, timedelta
 import time
+from datetime import datetime, timedelta
 
+from google.api_core import protobuf_helpers
 from google.cloud import spanner
+from google.cloud.exceptions import NotFound
 
 
 # [START spanner_create_backup]
 def create_backup(instance_id, database_id, backup_id, version_time):
     """Creates a backup for a database."""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
     # Create a backup
     expire_time = datetime.utcnow() + timedelta(days=14)
-    backup = instance.backup(
-        backup_id, database=database, expire_time=expire_time, version_time=version_time
+
+    request = backup_pb.CreateBackupRequest(
+        parent=instance.name,
+        backup_id=backup_id,
+        backup=backup_pb.Backup(
+            database=database.name,
+            expire_time=expire_time,
+            version_time=version_time,
+        ),
     )
-    operation = backup.create()
+
+    operation = spanner_client.database_admin_api.create_backup(request)
 
     # Wait for backup operation to complete.
-    operation.result(2100)
+    backup = operation.result(2100)
 
     # Verify that the backup is ready.
-    backup.reload()
-    assert backup.is_ready() is True
+    assert backup.state == backup_pb.Backup.State.READY
 
-    # Get the name, create time and backup size.
-    backup.reload()
     print(
         "Backup {} of size {} bytes was created at {} for version of database at {}".format(
             backup.name, backup.size_bytes, backup.create_time, backup.version_time
@@ -57,12 +68,17 @@ def create_backup(instance_id, database_id, backup_id, version_time):
 
 # [END spanner_create_backup]
 
+
 # [START spanner_create_backup_with_encryption_key]
 def create_backup_with_encryption_key(
     instance_id, database_id, backup_id, kms_key_name
 ):
     """Creates a backup for a database using a Customer Managed Encryption Key (CMEK)."""
-    from google.cloud.spanner_admin_database_v1 import CreateBackupEncryptionConfig
+
+    from google.cloud.spanner_admin_database_v1 import \
+        CreateBackupEncryptionConfig
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
 
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
@@ -74,23 +90,24 @@ def create_backup_with_encryption_key(
         "encryption_type": CreateBackupEncryptionConfig.EncryptionType.CUSTOMER_MANAGED_ENCRYPTION,
         "kms_key_name": kms_key_name,
     }
-    backup = instance.backup(
-        backup_id,
-        database=database,
-        expire_time=expire_time,
+    request = backup_pb.CreateBackupRequest(
+        parent=instance.name,
+        backup_id=backup_id,
+        backup=backup_pb.Backup(
+            database=database.name,
+            expire_time=expire_time,
+        ),
         encryption_config=encryption_config,
     )
-    operation = backup.create()
+    operation = spanner_client.database_admin_api.create_backup(request)
 
     # Wait for backup operation to complete.
-    operation.result(2100)
+    backup = operation.result(2100)
 
     # Verify that the backup is ready.
-    backup.reload()
-    assert backup.is_ready() is True
+    assert backup.state == backup_pb.Backup.State.READY
 
     # Get the name, create time, backup size and encryption key.
-    backup.reload()
     print(
         "Backup {} of size {} bytes was created at {} using encryption key {}".format(
             backup.name, backup.size_bytes, backup.create_time, kms_key_name
@@ -104,21 +121,24 @@ def create_backup_with_encryption_key(
 # [START spanner_restore_backup]
 def restore_database(instance_id, new_database_id, backup_id):
     """Restores a database from a backup."""
+    from google.cloud.spanner_admin_database_v1 import RestoreDatabaseRequest
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
-    # Create a backup on database_id.
 
     # Start restoring an existing backup to a new database.
-    backup = instance.backup(backup_id)
-    new_database = instance.database(new_database_id)
-    operation = new_database.restore(backup)
+    request = RestoreDatabaseRequest(
+        parent=instance.name,
+        database_id=new_database_id,
+        backup="{}/backups/{}".format(instance.name, backup_id),
+    )
+    operation = spanner_client.database_admin_api.restore_database(request)
 
     # Wait for restore operation to complete.
-    operation.result(1600)
+    db = operation.result(1600)
 
     # Newly created database has restore information.
-    new_database.reload()
-    restore_info = new_database.restore_info
+    restore_info = db.restore_info
     print(
         "Database {} restored to {} from backup {} with version time {}.".format(
             restore_info.backup_info.source_database,
@@ -137,34 +157,37 @@ def restore_database_with_encryption_key(
     instance_id, new_database_id, backup_id, kms_key_name
 ):
     """Restores a database from a backup using a Customer Managed Encryption Key (CMEK)."""
-    from google.cloud.spanner_admin_database_v1 import RestoreDatabaseEncryptionConfig
+    from google.cloud.spanner_admin_database_v1 import (
+        RestoreDatabaseEncryptionConfig, RestoreDatabaseRequest)
 
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
 
     # Start restoring an existing backup to a new database.
-    backup = instance.backup(backup_id)
     encryption_config = {
         "encryption_type": RestoreDatabaseEncryptionConfig.EncryptionType.CUSTOMER_MANAGED_ENCRYPTION,
         "kms_key_name": kms_key_name,
     }
-    new_database = instance.database(
-        new_database_id, encryption_config=encryption_config
+
+    request = RestoreDatabaseRequest(
+        parent=instance.name,
+        database_id=new_database_id,
+        backup="{}/backups/{}".format(instance.name, backup_id),
+        encryption_config=encryption_config,
     )
-    operation = new_database.restore(backup)
+    operation = spanner_client.database_admin_api.restore_database(request)
 
     # Wait for restore operation to complete.
-    operation.result(1600)
+    db = operation.result(1600)
 
     # Newly created database has restore information.
-    new_database.reload()
-    restore_info = new_database.restore_info
+    restore_info = db.restore_info
     print(
         "Database {} restored to {} from backup {} with using encryption key {}.".format(
             restore_info.backup_info.source_database,
             new_database_id,
             restore_info.backup_info.backup,
-            new_database.encryption_config.kms_key_name,
+            db.encryption_config.kms_key_name,
         )
     )
 
@@ -174,6 +197,9 @@ def restore_database_with_encryption_key(
 
 # [START spanner_cancel_backup_create]
 def cancel_backup(instance_id, database_id, backup_id):
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
@@ -181,9 +207,16 @@ def cancel_backup(instance_id, database_id, backup_id):
     expire_time = datetime.utcnow() + timedelta(days=30)
 
     # Create a backup.
-    backup = instance.backup(backup_id, database=database, expire_time=expire_time)
-    operation = backup.create()
+    request = backup_pb.CreateBackupRequest(
+        parent=instance.name,
+        backup_id=backup_id,
+        backup=backup_pb.Backup(
+            database=database.name,
+            expire_time=expire_time,
+        ),
+    )
 
+    operation = spanner_client.database_admin_api.create_backup(request)
     # Cancel backup creation.
     operation.cancel()
 
@@ -192,13 +225,22 @@ def cancel_backup(instance_id, database_id, backup_id):
     while not operation.done():
         time.sleep(300)  # 5 mins
 
-    # Deal with resource if the operation succeeded.
-    if backup.exists():
-        print("Backup was created before the cancel completed.")
-        backup.delete()
-        print("Backup deleted.")
-    else:
+    try:
+        spanner_client.database_admin_api.get_backup(
+            backup_pb.GetBackupRequest(
+                name="{}/backups/{}".format(instance.name, backup_id)
+            )
+        )
+    except NotFound:
         print("Backup creation was successfully cancelled.")
+        return
+    print("Backup was created before the cancel completed.")
+    spanner_client.database_admin_api.delete_backup(
+        backup_pb.DeleteBackupRequest(
+            name="{}/backups/{}".format(instance.name, backup_id)
+        )
+    )
+    print("Backup deleted.")
 
 
 # [END spanner_cancel_backup_create]
@@ -206,6 +248,9 @@ def cancel_backup(instance_id, database_id, backup_id):
 
 # [START spanner_list_backup_operations]
 def list_backup_operations(instance_id, database_id, backup_id):
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
 
@@ -215,9 +260,14 @@ def list_backup_operations(instance_id, database_id, backup_id):
         "google.spanner.admin.database.v1.CreateBackupMetadata) "
         "AND (metadata.database:{})"
     ).format(database_id)
-    operations = instance.list_backup_operations(filter_=filter_)
+    request = backup_pb.ListBackupOperationsRequest(
+        parent=instance.name, filter=filter_
+    )
+    operations = spanner_client.database_admin_api.list_backup_operations(request)
     for op in operations:
-        metadata = op.metadata
+        metadata = protobuf_helpers.from_any_pb(
+            backup_pb.CreateBackupMetadata, op.metadata
+        )
         print(
             "Backup {} on database {}: {}% complete.".format(
                 metadata.name, metadata.database, metadata.progress.progress_percent
@@ -229,9 +279,14 @@ def list_backup_operations(instance_id, database_id, backup_id):
         "(metadata.@type:type.googleapis.com/google.spanner.admin.database.v1.CopyBackupMetadata) "
         "AND (metadata.source_backup:{})"
     ).format(backup_id)
-    operations = instance.list_backup_operations(filter_=filter_)
+    request = backup_pb.ListBackupOperationsRequest(
+        parent=instance.name, filter=filter_
+    )
+    operations = spanner_client.database_admin_api.list_backup_operations(request)
     for op in operations:
-        metadata = op.metadata
+        metadata = protobuf_helpers.from_any_pb(
+            backup_pb.CopyBackupMetadata, op.metadata
+        )
         print(
             "Backup {} on source backup {}: {}% complete.".format(
                 metadata.name,
@@ -246,6 +301,9 @@ def list_backup_operations(instance_id, database_id, backup_id):
 
 # [START spanner_list_database_operations]
 def list_database_operations(instance_id):
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
 
@@ -254,36 +312,100 @@ def list_database_operations(instance_id):
         "(metadata.@type:type.googleapis.com/"
         "google.spanner.admin.database.v1.OptimizeRestoredDatabaseMetadata)"
     )
-    operations = instance.list_database_operations(filter_=filter_)
+    request = spanner_database_admin.ListDatabaseOperationsRequest(
+        parent=instance.name, filter=filter_
+    )
+    operations = spanner_client.database_admin_api.list_database_operations(request)
     for op in operations:
+        metadata = protobuf_helpers.from_any_pb(
+            spanner_database_admin.OptimizeRestoredDatabaseMetadata, op.metadata
+        )
         print(
             "Database {} restored from backup is {}% optimized.".format(
-                op.metadata.name, op.metadata.progress.progress_percent
+                metadata.name, metadata.progress.progress_percent
             )
         )
 
 
 # [END spanner_list_database_operations]
+# [START spanner_postgresql_jsonb_add_column]
+def add_jsonb_column(instance_id, database_id):
+    """
+    Alters Venues tables in the database adding a JSONB column.
+    You can create the table by running the `create_table_with_datatypes`
+    sample or by running this DDL statement against your database:
+    CREATE TABLE Venues (
+      VenueId         BIGINT NOT NULL,
+      VenueName       character varying(100),
+      VenueInfo       BYTEA,
+      Capacity        BIGINT,
+      OutdoorVenue    BOOL,
+      PopularityScore FLOAT8,
+      Revenue         NUMERIC,
+      LastUpdateTime  SPANNER.COMMIT_TIMESTAMP NOT NULL,
+      PRIMARY KEY (VenueId))
+    """
+    # instance_id = "your-spanner-instance"
+    # database_id = "your-spanner-db-id"
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
+    spanner_client = spanner.Client()
+    instance = spanner_client.instance(instance_id)
+    database = instance.database(database_id)
+
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
+        database=database.name,
+        statements=["ALTER TABLE Venues ADD COLUMN VenueDetails JSONB"],
+    )
+
+    operation = spanner_client.database_admin_api.update_database_ddl(request)
+
+    print("Waiting for operation to complete...")
+    operation.result(OPERATION_TIMEOUT_SECONDS)
+
+    print(
+        'Altered table "Venues" on database {} on instance {}.'.format(
+            database_id, instance_id
+        )
+    )
+
+
+# [END spanner_postgresql_jsonb_add_column]
 
 
 # [START spanner_list_backups]
 def list_backups(instance_id, database_id, backup_id):
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
 
     # List all backups.
     print("All backups:")
-    for backup in instance.list_backups():
+    request = backup_pb.ListBackupsRequest(parent=instance.name, filter="")
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         print(backup.name)
 
     # List all backups that contain a name.
     print('All backups with backup name containing "{}":'.format(backup_id))
-    for backup in instance.list_backups(filter_="name:{}".format(backup_id)):
+    request = backup_pb.ListBackupsRequest(
+        parent=instance.name, filter="name:{}".format(backup_id)
+    )
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         print(backup.name)
 
     # List all backups for a database that contains a name.
     print('All backups with database name containing "{}":'.format(database_id))
-    for backup in instance.list_backups(filter_="database:{}".format(database_id)):
+    request = backup_pb.ListBackupsRequest(
+        parent=instance.name, filter="database:{}".format(database_id)
+    )
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         print(backup.name)
 
     # List all backups that expire before a timestamp.
@@ -293,14 +415,21 @@ def list_backups(instance_id, database_id, backup_id):
             *expire_time.timetuple()
         )
     )
-    for backup in instance.list_backups(
-        filter_='expire_time < "{}-{}-{}T{}:{}:{}Z"'.format(*expire_time.timetuple())
-    ):
+    request = backup_pb.ListBackupsRequest(
+        parent=instance.name,
+        filter='expire_time < "{}-{}-{}T{}:{}:{}Z"'.format(*expire_time.timetuple()),
+    )
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         print(backup.name)
 
     # List all backups with a size greater than some bytes.
     print("All backups with backup size more than 100 bytes:")
-    for backup in instance.list_backups(filter_="size_bytes > 100"):
+    request = backup_pb.ListBackupsRequest(
+        parent=instance.name, filter="size_bytes > 100"
+    )
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         print(backup.name)
 
     # List backups that were created after a timestamp that are also ready.
@@ -310,18 +439,23 @@ def list_backups(instance_id, database_id, backup_id):
             *create_time.timetuple()
         )
     )
-    for backup in instance.list_backups(
-        filter_='create_time >= "{}-{}-{}T{}:{}:{}Z" AND state:READY'.format(
+    request = backup_pb.ListBackupsRequest(
+        parent=instance.name,
+        filter='create_time >= "{}-{}-{}T{}:{}:{}Z" AND state:READY'.format(
             *create_time.timetuple()
-        )
-    ):
+        ),
+    )
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         print(backup.name)
 
     print("All backups with pagination")
     # If there are multiple pages, additional ``ListBackup``
     # requests will be made as needed while iterating.
     paged_backups = set()
-    for backup in instance.list_backups(page_size=2):
+    request = backup_pb.ListBackupsRequest(parent=instance.name, page_size=2)
+    operations = spanner_client.database_admin_api.list_backups(request)
+    for backup in operations:
         paged_backups.add(backup.name)
     for backup in paged_backups:
         print(backup)
@@ -332,22 +466,39 @@ def list_backups(instance_id, database_id, backup_id):
 
 # [START spanner_delete_backup]
 def delete_backup(instance_id, backup_id):
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
-    backup = instance.backup(backup_id)
-    backup.reload()
+    backup = spanner_client.database_admin_api.get_backup(
+        backup_pb.GetBackupRequest(
+            name="{}/backups/{}".format(instance.name, backup_id)
+        )
+    )
 
     # Wait for databases that reference this backup to finish optimizing.
     while backup.referencing_databases:
         time.sleep(30)
-        backup.reload()
+        backup = spanner_client.database_admin_api.get_backup(
+            backup_pb.GetBackupRequest(
+                name="{}/backups/{}".format(instance.name, backup_id)
+            )
+        )
 
     # Delete the backup.
-    backup.delete()
+    spanner_client.database_admin_api.delete_backup(
+        backup_pb.DeleteBackupRequest(name=backup.name)
+    )
 
     # Verify that the backup is deleted.
-    assert backup.exists() is False
-    print("Backup {} has been deleted.".format(backup.name))
+    try:
+        backup = spanner_client.database_admin_api.get_backup(
+            backup_pb.GetBackupRequest(name=backup.name)
+        )
+    except NotFound:
+        print("Backup {} has been deleted.".format(backup.name))
+        return
 
 
 # [END spanner_delete_backup]
@@ -355,16 +506,28 @@ def delete_backup(instance_id, backup_id):
 
 # [START spanner_update_backup]
 def update_backup(instance_id, backup_id):
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
-    backup = instance.backup(backup_id)
-    backup.reload()
+
+    backup = spanner_client.database_admin_api.get_backup(
+        backup_pb.GetBackupRequest(
+            name="{}/backups/{}".format(instance.name, backup_id)
+        )
+    )
 
     # Expire time must be within 366 days of the create time of the backup.
     old_expire_time = backup.expire_time
     # New expire time should be less than the max expire time
     new_expire_time = min(backup.max_expire_time, old_expire_time + timedelta(days=30))
-    backup.update_expire_time(new_expire_time)
+    spanner_client.database_admin_api.update_backup(
+        backup_pb.UpdateBackupRequest(
+            backup=backup_pb.Backup(name=backup.name, expire_time=new_expire_time),
+            update_mask={"paths": ["expire_time"]},
+        )
+    )
     print(
         "Backup {} expire time was updated from {} to {}.".format(
             backup.name, old_expire_time, new_expire_time
@@ -380,6 +543,10 @@ def create_database_with_version_retention_period(
     instance_id, database_id, retention_period
 ):
     """Creates a database with a version retention period."""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     ddl_statements = [
@@ -400,20 +567,24 @@ def create_database_with_version_retention_period(
             database_id, retention_period
         ),
     ]
-    db = instance.database(database_id, ddl_statements)
-    operation = db.create()
-
-    operation.result(30)
-
-    db.reload()
-
-    print(
-        "Database {} created with version retention period {} and earliest version time {}".format(
-            db.database_id, db.version_retention_period, db.earliest_version_time
+    operation = spanner_client.database_admin_api.create_database(
+        request=spanner_database_admin.CreateDatabaseRequest(
+            parent=instance.name,
+            create_statement="CREATE DATABASE `{}`".format(database_id),
+            extra_statements=ddl_statements,
         )
     )
 
-    db.drop()
+    db = operation.result(30)
+    print(
+        "Database {} created with version retention period {} and earliest version time {}".format(
+            db.name, db.version_retention_period, db.earliest_version_time
+        )
+    )
+
+    spanner_client.database_admin_api.drop_database(
+        spanner_database_admin.DropDatabaseRequest(database=db.name)
+    )
 
 
 # [END spanner_create_database_with_version_retention_period]
@@ -422,22 +593,29 @@ def create_database_with_version_retention_period(
 # [START spanner_copy_backup]
 def copy_backup(instance_id, backup_id, source_backup_path):
     """Copies a backup."""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        backup as backup_pb
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
 
     # Create a backup object and wait for copy backup operation to complete.
     expire_time = datetime.utcnow() + timedelta(days=14)
-    copy_backup = instance.copy_backup(
-        backup_id=backup_id, source_backup=source_backup_path, expire_time=expire_time
+    request = backup_pb.CopyBackupRequest(
+        parent=instance.name,
+        backup_id=backup_id,
+        source_backup=source_backup_path,
+        expire_time=expire_time,
     )
-    operation = copy_backup.create()
 
-    # Wait for copy backup operation to complete.
-    operation.result(2100)
+    operation = spanner_client.database_admin_api.copy_backup(request)
+
+    # Wait for backup operation to complete.
+    copy_backup = operation.result(2100)
 
     # Verify that the copy backup is ready.
-    copy_backup.reload()
-    assert copy_backup.is_ready() is True
+    assert copy_backup.state == backup_pb.Backup.State.READY
 
     print(
         "Backup {} of size {} bytes was created at {} with version time {}".format(
