+++ samples/samples/pg_snippets.py
@@ -68,26 +68,34 @@ def create_instance(instance_id):
 # [START spanner_postgresql_create_database]
 def create_database(instance_id, database_id):
     """Creates a PostgreSql database and tables for sample data."""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
 
-    database = instance.database(
-        database_id,
+    request = spanner_database_admin.CreateDatabaseRequest(
+        parent=instance.name,
+        create_statement=f'CREATE DATABASE "{database_id}"',
         database_dialect=DatabaseDialect.POSTGRESQL,
     )
 
-    operation = database.create()
+    operation = spanner_client.database_admin_api.create_database(request=request)
 
     print("Waiting for operation to complete...")
-    operation.result(OPERATION_TIMEOUT_SECONDS)
+    database = operation.result(OPERATION_TIMEOUT_SECONDS)
 
     create_table_using_ddl(database.name)
     print("Created database {} on instance {}".format(database_id, instance_id))
 
 
 def create_table_using_ddl(database_name):
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
-    request = spanner_admin_database_v1.UpdateDatabaseDdlRequest(
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
         database=database_name,
         statements=[
             """CREATE TABLE Singers (
@@ -231,13 +239,19 @@ def read_data(instance_id, database_id):
 # [START spanner_postgresql_add_column]
 def add_column(instance_id, database_id):
     """Adds a new column to the Albums table in the example database."""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    operation = database.update_ddl(
-        ["ALTER TABLE Albums ADD COLUMN MarketingBudget BIGINT"]
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
+        database=database.name,
+        statements=["ALTER TABLE Albums ADD COLUMN MarketingBudget BIGINT"],
     )
+    operation = spanner_client.database_admin_api.update_database_ddl(request)
 
     print("Waiting for operation to complete...")
     operation.result(OPERATION_TIMEOUT_SECONDS)
@@ -390,6 +404,7 @@ def add_index(instance_id, database_id):
 
 # [END spanner_postgresql_create_index]
 
+
 # [START spanner_postgresql_read_data_with_index]
 def read_data_with_index(instance_id, database_id):
     """Reads sample data from the database using an index.
@@ -424,16 +439,23 @@ def read_data_with_index(instance_id, database_id):
 # [START spanner_postgresql_create_storing_index]
 def add_storing_index(instance_id, database_id):
     """Adds an storing index to the example database."""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    operation = database.update_ddl(
-        [
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
+        database=database.name,
+        statements=[
             "CREATE INDEX AlbumsByAlbumTitle2 ON Albums(AlbumTitle)"
             "INCLUDE (MarketingBudget)"
-        ]
+        ],
     )
+
+    operation = spanner_client.database_admin_api.update_database_ddl(request)
 
     print("Waiting for operation to complete...")
     operation.result(OPERATION_TIMEOUT_SECONDS)
@@ -1066,11 +1088,15 @@ def create_table_with_datatypes(instance_id, database_id):
     # [START spanner_postgresql_create_table_with_datatypes]
     # instance_id = "your-spanner-instance"
     # database_id = "your-spanner-db-id"
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    request = spanner_admin_database_v1.UpdateDatabaseDdlRequest(
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
         database=database.name,
         statements=[
             """CREATE TABLE Venues (
@@ -1447,13 +1473,19 @@ def add_jsonb_column(instance_id, database_id):
     # instance_id = "your-spanner-instance"
     # database_id = "your-spanner-db-id"
 
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    operation = database.update_ddl(
-        ["ALTER TABLE Venues ADD COLUMN VenueDetails JSONB"]
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
+        database=database.name,
+        statements=["ALTER TABLE Venues ADD COLUMN VenueDetails JSONB"],
     )
+
+    operation = spanner_client.database_admin_api.update_database_ddl(request)
 
     print("Waiting for operation to complete...")
     operation.result(OPERATION_TIMEOUT_SECONDS)
@@ -1524,6 +1556,7 @@ def update_data_with_jsonb(instance_id, database_id):
 
 # [END spanner_postgresql_jsonb_update_data]
 
+
 # [START spanner_postgresql_jsonb_query_parameter]
 def query_data_with_jsonb_parameter(instance_id, database_id):
     """Queries sample data using SQL with a JSONB parameter."""
@@ -1555,11 +1588,15 @@ def query_data_with_jsonb_parameter(instance_id, database_id):
 # [START spanner_postgresql_create_sequence]
 def create_sequence(instance_id, database_id):
     """Creates the Sequence and insert data"""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    request = spanner_admin_database_v1.UpdateDatabaseDdlRequest(
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
         database=database.name,
         statements=[
             "CREATE SEQUENCE Seq BIT_REVERSED_POSITIVE",
@@ -1601,14 +1638,23 @@ def create_sequence(instance_id, database_id):
 
 # [END spanner_postgresql_create_sequence]
 
+
 # [START spanner_postgresql_alter_sequence]
 def alter_sequence(instance_id, database_id):
     """Alters the Sequence and insert data"""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    operation = database.update_ddl(["ALTER SEQUENCE Seq SKIP RANGE 1000 5000000"])
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
+        database=database.name,
+        statements=["ALTER SEQUENCE Seq SKIP RANGE 1000 5000000"],
+    )
+    operation = spanner_client.database_admin_api.update_database_ddl(request)
 
     print("Waiting for operation to complete...")
     operation.result(OPERATION_TIMEOUT_SECONDS)
@@ -1640,19 +1686,26 @@ def alter_sequence(instance_id, database_id):
 
 # [END spanner_postgresql_alter_sequence]
 
+
 # [START spanner_postgresql_drop_sequence]
 def drop_sequence(instance_id, database_id):
     """Drops the Sequence"""
+
+    from google.cloud.spanner_admin_database_v1.types import \
+        spanner_database_admin
+
     spanner_client = spanner.Client()
     instance = spanner_client.instance(instance_id)
     database = instance.database(database_id)
 
-    operation = database.update_ddl(
-        [
+    request = spanner_database_admin.UpdateDatabaseDdlRequest(
+        database=database.name,
+        statements=[
             "ALTER TABLE Customers ALTER COLUMN CustomerId DROP DEFAULT",
             "DROP SEQUENCE Seq",
-        ]
+        ],
     )
+    operation = spanner_client.database_admin_api.update_database_ddl(request)
 
     print("Waiting for operation to complete...")
     operation.result(OPERATION_TIMEOUT_SECONDS)
