+++ samples/samples/pg_snippets_test.py
@@ -15,9 +15,9 @@
 import time
 import uuid
 
+import pytest
 from google.api_core import exceptions
 from google.cloud.spanner_admin_database_v1.types.common import DatabaseDialect
-import pytest
 from test_utils.retry import RetryErrors
 
 import pg_snippets as snippets
